✨ [feat] : 채팅방 aside의 채팅방 리스트 컴포넌트와 채팅방 컴포넌트 분리
